Remove Test Header in linkprediction
